==> scripts/01_Prep_Dem-Stg-ADI_Data_for_Regression.Rmd <==
---
title: "Prepare Demog-Staging-ADI Data for Regression"
author: "Steph Reynolds (Stephanie.Reynolds@ucsf.edu)"
date: "`r format(Sys.time(), '%b %d, %Y  %H:%M %p')`"
output:
  html_document:
    toc: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Background

## Project 
MINDSCAPE: Modeling of infectious network dynamics for surveillance, control and prevention enhancement

## Description
This file cleans the demographics, Covid staging, and ADI data (`dm_covid_all_7_19_21.csv`) and prepares the dataset for regression analysis.

# Load required packages
```{r Load required packages, include=F}
library(tidyverse)
```

# Import and preview data; keep only necessary variables 
```{r Import and preview data}
raw <- read_delim("/Users/sreynolds2/Documents/GitHub/MS-ADI/data/raw/dm_covid_all_7_19_21.csv", delim = "|")

# View colnames and identify variables to keep and assign to vector
colnames(raw)
vars <- c("deid_enc_id", "age", "sex", "smoking", "BMI", "cci_total", "marital_status", "language", "insurance_type", "pat_race", "ethnicity", "ADI_STATERNK_y", "max_stage", "readmit", "hospital_LOS", "end_in_death", "admission_stage")

df <- raw %>% 
  filter(!is.na(ADI_STATERNK_y)) %>%   # Remove rows with missing ADI value
  filter(hospital_LOS>=1) %>%          # Only include rows where LOS>=1
  select(vars) %>%                     # Keep only variables in `vars` vector
  rename(ADI = ADI_STATERNK_y)         # Rename ADI variable for convenience 

n_distinct(df) # 825 unique patients included in final dataset 
```

# Recode variables, and collapse levels for factor variables 
```{r Recode variables}
# Identify variables to be converted to factor type and assign to vector `vars_fct`
vars_fct <- c("sex", "smoking", "marital_status", "insurance_type", "pat_race", "ethnicity")

# Transform variables above to factor type
df <- df %>% mutate_at(vars_fct, factor)

# Collapse levels for race
df$pat_race <- fct_collapse(df$pat_race,
                            "Unknown/Declined" = c("Unknown", "Declined", "Unknown/Declined"),
                            "Other" = c("Native Hawaiian or Other Pacific Islander", "Other", "Other Pacific Islander", "American Indian or Alaska Native")) %>% 
  relevel(df$pat_race, ref = "White or Caucasian")

# Collapse levels for ethnicity
df$ethnicity <- fct_collapse(df$ethnicity, 
                             "Unknown/Declined" = c("Unknown", "Declined", "Unknown/Declined")) %>% 
  relevel(df$ethnicity, ref = "Not Hispanic or Latino")

# Collapse levels for insurance type 
df$insurance_type <- fct_collapse(df$insurance_type, 
                                  "Medicare/Medi-Cal" = c("01 - Medicare", "02 - Medi-Cal"),
                                  "Private Insurance" = c("03 - Private Coverage", "04 - Workers Compensation", "06 - Other Government")) %>% 
  relevel(df$insurance_type, ref = "Private Insurance")

# Collapse levels for smoking
df$smoking <- fct_collapse(df$smoking,
                           "Current Smoker" = c("Current Every Day Smoker", "Current Some Day Smoker", "Light Tobacco Smoker", "Smoker, Current Status Unknown"),
                           "Not Smoker" = c("Former Smoker", "Never Smoker", "Passive Smoke Exposure - Never Smoker"),
                           "Unknown" = "Unknown If Ever Smoked") %>% 
  relevel(df$smoking, ref = "Not Smoker")

# Collapse levels for marital status
df$marital_status <- fct_collapse(df$marital_status,
                                  "Single" = c("Divorced", "Legally Separated", "RDP-Dissolved", "Single", "Widowed"), 
                                  "Paired" = c("Married", "Significant Other")) %>% 
  relevel(df$marital_status, ref = "Single")

# Collapse levels for language 
df$language <- factor(ifelse(df$language=="English", "English", "Not English")) %>% 
  relevel(df$language, ref = "English")

# Recode `end_in_death` as binary variable where yes=1
df$end_in_death <- as.integer(df$end_in_death=="Yes", 1, 0)

# Recode `readmit` as binary variable where yes=1
df$readmit <- as.integer(df$readmit=="Yes", 1, 0)

# Create `severe_stg` binary variable to indicate whether patient had a max stage of 6 or above (stages 6-9 considered severe by WHO clinical stages, and stage 10 considered dead)
df$severe_stg <- ifelse(df$max_stage>=7, 1, 0)

# Recode `admission_stage` where stages 7-9 are grouped into 7; leave stages 4-6 as-is
df$admission_stage <- case_when(df$admission_stage==4 ~ 4,
                                df$admission_stage==5 ~ 5,
                                df$admission_stage==6 ~ 6,
                                df$admission_stage %in% 7:9 ~ 7)

# Set reference level for `sex` 
df$sex <- relevel(df$sex, ref = "Male")
```

# Save dataframe as CSV and RSD
```{r Save df as RSD, eval = FALSE}
saveRDS(df, "/Users/sreynolds2/Documents/GitHub/MS-ADI/data/dm_stg_adi_7.19.21.rds")
write_csv(df, "/Users/sreynolds2/Documents/GitHub/MS-ADI/data/dm_stg_adi_7.19.21.csv")
```
