==> scripts/02_Regression_Dem-ADI-Staging_Data_SMM.Rmd <==
---
title: "Regression Analysis for Demog-Staging-ADI Data -- Primary Outcome: Severe Morbidity and Mortality (SMM)"
author: "Steph Reynolds (Stephanie.Reynolds@ucsf.edu)"
date: "`r format(Sys.time(), '%b %d, %Y  %H:%M %p')`"
output:
  html_document:
    toc: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Background

## Project

MINDSCAPE: Modeling of infectious network dynamics for surveillance, control and prevention enhancement

## Description

This file reads in the demographics, Covid staging, and ADI data (`dm_stg_adi_7.19.21.rds`) and fits a logistic regression model with severe morbidity and mortality (`severe_stg`) as the outcome. Primary and secondary outcomes for full regression analysis can be found below.

## Primary Outcome

-   Severe Morbidity and Mortality (`severe_stg`) -\> All patients who reached a max stage of 6+ (severe) were recoded as severe_stg = 1

## Secondary Outcomes

-   Length of Stay (`LOS`) -\> Length of stay that patients were hospitalized
    -   Only patients who were hospitalized for \>= 1 day - aka inpatients - were included in the regression analysis.
-   Readmit (`readmit`) --\> Patients who were readmitted to hospital within 30(?) days

# Load required packages

```{r Load required packages, include=F}
library(tidyverse)
library(tidymodels)
library(performance) # to calculate model performance and check model assumptions
```

# Import and preview data

```{r Read in df}
df <- readRDS("/Users/sreynolds2/Documents/GitHub/MS-ADI/data/dm_stg_adi_7.19.21.rds")

head(df, n=10)
```

# Compare full and null models for severe morbidity and mortality (SMM)

```{r Full and null models of SMM}
# FULL MODEL 
m_full <- glm(severe_stg ~ age + sex + smoking + cci_total + marital_status + language + insurance_type + pat_race + ethnicity + ADI + admission_stage + BMI, data = df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_full), confint.default(m_full, level=0.95)))

# Summarize regression output 
summary(m_full)

# Check model assumptions
check_model(m_full)

# Check model performance 
model_performance(m_full)

# NULL MODEL 
m_null <- glm(severe_stg ~ 1, data = df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_null), confint.default(m_null, level=0.95)))

# Summarize regression output 
summary(m_null)

# Check model performance 
model_performance(m_null)
```

# Generate simple logistic regressions for severe morbidity and mortality

## Severe_stg \~ age

```{r Severe_stg ~ age}
m_sev_age <- glm(severe_stg ~ age, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_age), confint.default(m_sev_age, level=0.95)))

# Summarize regression output 
summary(m_sev_age)

# Check model assumptions
check_model(m_sev_age)

# Check model performance 
model_performance(m_sev_age)
```

## Severe_stg \~ sex

```{r Severe_stg ~ sex}
m_sev_sex <- glm(severe_stg ~ sex, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_sex), confint.default(m_sev_sex, level=0.95)))

# Summarize regression output 
summary(m_sev_sex)

# Check model assumptions
check_model(m_sev_sex)

# Check model performance 
model_performance(m_sev_sex)
```

## Severe_stg \~ smoking

```{r Severe_stg ~ smoking}
m_sev_smoke <- glm(severe_stg ~ smoking, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_smoke), confint.default(m_sev_smoke, level=0.95)))

# Summarize regression output 
summary(m_sev_smoke)

# Check model assumptions
check_model(m_sev_smoke)

# Check model performance 
model_performance(m_sev_smoke)
```

## Severe_stg \~ Charlston Comorbidity Index (CCI)

```{r Severe_stg ~ cci}
m_sev_cci <- glm(severe_stg ~ cci_total, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_cci), confint.default(m_sev_cci, level=0.95)))

# Summarize regression output 
summary(m_sev_cci)

# Check model assumptions
check_model(m_sev_cci)

# Check model performance 
model_performance(m_sev_cci)
```

## Severe_stg \~ marital status

```{r Severe_stg ~ marital_status}
m_sev_marital <- glm(severe_stg ~ marital_status, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_marital), confint.default(m_sev_marital, level=0.95)))

# Summarize regression output 
summary(m_sev_marital)

# Check model assumptions
check_model(m_sev_marital)

# Check model performance 
model_performance(m_sev_marital)
```

## Severe_stg \~ language

```{r Severe_stg ~ language}
m_sev_language <- glm(severe_stg ~ language, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_language), confint.default(m_sev_language, level=0.95)))

# Summarize regression output 
summary(m_sev_language)

# Check model assumptions
check_model(m_sev_language)

# Check model performance 
model_performance(m_sev_language)
```

## Severe_stg \~ insurance type

```{r Severe_stg ~ insurance_type}
m_sev_insur <- glm(severe_stg ~ insurance_type, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_insur), confint.default(m_sev_insur, level=0.95)))

# Summarize regression output 
summary(m_sev_insur)

# Check model assumptions
check_model(m_sev_insur)

# Check model performance 
model_performance(m_sev_insur)
```

## Severe_stg \~ race

```{r Severe_stg ~ race}
m_sev_race <- glm(severe_stg ~ pat_race, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_race), confint.default(m_sev_race, level=0.95)))

# Summarize regression output 
summary(m_sev_race)

# Check model assumptions
check_model(m_sev_race)

# Check model performance 
model_performance(m_sev_race)
```

## Severe_stg \~ ethnicity

```{r Severe_stg ~ ethnicity}
m_sev_ethnicity <- glm(severe_stg ~ ethnicity, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_ethnicity), confint.default(m_sev_ethnicity, level=0.95)))

# Summarize regression output 
summary(m_sev_ethnicity)

# Check model assumptions
check_model(m_sev_ethnicity)

# Check model performance 
model_performance(m_sev_ethnicity)
```

## Severe_stg \~ Area Deprivation Index (ADI)

```{r Severe_stg ~ adi}
m_sev_adi <- glm(severe_stg ~ ADI, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_adi), confint.default(m_sev_adi, level=0.95)))

# Summarize regression output 
summary(m_sev_adi)

# Check model assumptions
check_model(m_sev_adi)

# Check model performance 
model_performance(m_sev_adi)
```

## Severe_stg \~ admission stage

```{r Severe_stg ~ admission_stage}
m_sev_adm_stg <- glm(severe_stg ~ admission_stage, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_adm_stg), confint.default(m_sev_adm_stg, level=0.95)))

# Summarize regression output 
summary(m_sev_adm_stg)

# Check model assumptions
check_model(m_sev_adm_stg)

# Check model performance 
model_performance(m_sev_adm_stg)
```

## Severe_stg \~ BMI

```{r Severe_stg ~ BMI}
m_sev_bmi <- glm(severe_stg ~ BMI, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_sev_bmi), confint.default(m_sev_bmi, level=0.95)))

# Summarize regression output 
summary(m_sev_bmi)

# Check model assumptions
check_model(m_sev_bmi)

# Check model performance 
model_performance(m_sev_bmi)
```
