==> scripts/03_Regression_Dem-ADI-Staging_Data_LOS.Rmd <==
---
title: "Regression Analysis for Demog-Staging-ADI Data -- Secondary Outcome: Length of Stay (LOS)"
author: "Steph Reynolds (Stephanie.Reynolds@ucsf.edu)"
date: "`r format(Sys.time(), '%b %d, %Y  %H:%M %p')`"
output:
  html_document:
    toc: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Background

## Project 
MINDSCAPE: Modeling of infectious network dynamics for surveillance, control and prevention enhancement

## Description
This file reads in the demographics, Covid staging, and ADI data (`dm_stg_adi_7.19.21.rds`) and fits a linear regression model with length of stay (`LOS`) as the outcome. Primary and secondary outcomes for full regression analysis can be found below. 

## Primary Outcome 

  * Morbidity and Mortality (`severe_stg`) -> All patients who reached a max stage of 6+ (severe) were recoded as severe_stg = 1

## Secondary Outcomes

  * Length of Stay (`LOS`) -> Length of stay that patients were hospitalized
    * Only patients who were hospitalized for >= 1 day - aka inpatients - were included in the regression analysis.
    * NOTE: Since LOS was non-normally distributed, we chose to transform it on the log scale. 
  * Readmit (`readmit`) --> Patients who were readmitted to hospital within 30(?) days

# Load required packages
```{r Load required packages, include=F}
library(tidyverse)
library(tidymodels)
library(performance) # to calculate model performance and check model assumptions
```

# Import and preview data
```{r Read in df}
df <- readRDS("/Users/sreynolds2/Documents/GitHub/MS-ADI/data/dm_stg_adi_7.19.21.rds") %>% 
  rename(LOS = hospital_LOS)     # rename for convenience 

head(df, n=10)
```

# Compare full and null models -- Length of stay (`LOS`)
```{r Full and null models for LOS}
# FULL MODEL 
m_full_los <- lm(log(LOS) ~ age + sex + smoking + cci_total + marital_status + language + insurance_type + pat_race + ethnicity + ADI + admission_stage + BMI, data = df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_full_los), confint.default(m_full_los, level=0.95)))

# Summarize regression output 
summary(m_full_los)

# Check model assumptions
check_model(m_full_los)

# Check model performance 
model_performance(m_full_los)

# NULL MODEL 
m_null_los <- lm(log(LOS) ~ 1, data = df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_null_los), confint.default(m_null_los, level=0.95)))

# Summarize regression output 
summary(m_null_los)

# Check model performance 
model_performance(m_null_los)
```

# Generate simple logistic regressions for length of stay (LOS)

## LOS ~ age
```{r LOS ~ age}
m_LOS_age <- lm(log(LOS) ~ age, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_age), confint.default(m_LOS_age, level=0.95)))

# Summarize regression output 
summary(m_LOS_age)

# Check model assumptions
check_model(m_LOS_age)

# Check model performance 
model_performance(m_LOS_age)
```

## LOS ~ sex 
```{r LOS ~ sex}
m_LOS_sex <- lm(log(LOS) ~ sex, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_sex), confint.default(m_LOS_sex, level=0.95)))

# Summarize regression output 
summary(m_LOS_sex)

# Check model assumptions
check_model(m_LOS_sex)

# Check model performance 
model_performance(m_LOS_sex)
```

## LOS ~ smoking 
```{r LOS ~ smoking}
m_LOS_smoking <- lm(log(LOS) ~ smoking, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_smoking), confint.default(m_LOS_smoking, level=0.95)))

# Summarize regression output 
summary(m_LOS_smoking)

# Check model assumptions
check_model(m_LOS_smoking)

# Check model performance 
model_performance(m_LOS_smoking)
```

## LOS ~ Charlston Comorbidity Index (CCI)
```{r LOS ~ CCI}
m_LOS_cci <- lm(log(LOS) ~ cci_total, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_cci), confint.default(m_LOS_cci, level=0.95)))

# Summarize regression output 
summary(m_LOS_cci)

# Check model assumptions
check_model(m_LOS_cci)

# Check model performance 
model_performance(m_LOS_cci)
```

## LOS ~ marital status 
```{r LOS ~ marital status}
m_LOS_marital <- lm(log(LOS) ~ marital_status, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_marital), confint.default(m_LOS_marital, level=0.95)))

# Summarize regression output 
summary(m_LOS_marital)

# Check model assumptions
check_model(m_LOS_marital)

# Check model performance 
model_performance(m_LOS_marital)
```

## LOS ~ language
```{r LOS ~ language}
m_LOS_language <- lm(log(LOS) ~ language, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_language), confint.default(m_LOS_language, level=0.95)))

# Summarize regression output 
summary(m_LOS_language)

# Check model assumptions
check_model(m_LOS_language)

# Check model performance 
model_performance(m_LOS_language)
```

## LOS ~ insurance type
```{r LOS ~ insurance type}
m_LOS_insur <- lm(log(LOS) ~ insurance_type, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_insur), confint.default(m_LOS_insur, level=0.95)))

# Summarize regression output 
summary(m_LOS_insur)

# Check model assumptions
check_model(m_LOS_insur)

# Check model performance 
model_performance(m_LOS_insur)
```

## LOS ~ race
```{r LOS ~ race}
m_LOS_race <- lm(log(LOS) ~ pat_race, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_race), confint.default(m_LOS_race, level=0.95)))

# Summarize regression output 
summary(m_LOS_race)

# Check model assumptions
check_model(m_LOS_race)

# Check model performance 
model_performance(m_LOS_race)
```

## LOS ~ ethnicity
```{r LOS ~ ethnicity}
m_LOS_ethnicity <- lm(log(LOS) ~ ethnicity, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_ethnicity), confint.default(m_LOS_ethnicity, level=0.95)))

# Summarize regression output 
summary(m_LOS_ethnicity)

# Check model assumptions
check_model(m_LOS_ethnicity)

# Check model performance 
model_performance(m_LOS_ethnicity)
```

## LOS ~ Area Deprivation Index (ADI)
```{r LOS ~ ADI}
m_LOS_ADI <- lm(log(LOS) ~ ADI, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_ADI), confint.default(m_LOS_ADI, level=0.95)))

# Summarize regression output 
summary(m_LOS_ADI)

# Check model assumptions
check_model(m_LOS_ADI)

# Check model performance 
model_performance(m_LOS_ADI)
```

## LOS ~ admission stage 
```{r LOS ~ admssion stage}
m_LOS_adm_stg <- lm(log(LOS) ~ admission_stage, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_adm_stg), confint.default(m_LOS_adm_stg, level=0.95)))

# Summarize regression output 
summary(m_LOS_adm_stg)

# Check model assumptions
check_model(m_LOS_adm_stg)

# Check model performance 
model_performance(m_LOS_adm_stg)
```

## LOS ~ BMI
```{r LOS ~ BMI}
m_LOS_BMI <- lm(log(LOS) ~ BMI, df)

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_LOS_BMI), confint.default(m_LOS_BMI, level=0.95)))

# Summarize regression output 
summary(m_LOS_BMI)

# Check model assumptions
check_model(m_LOS_BMI)

# Check model performance 
model_performance(m_LOS_BMI)
```
