==> scripts/04_Regression_Dem-ADI-Staging_Data_Readmit.Rmd <==
---
title: "Regression Analysis for Demog-Staging-ADI Data -- Secondary Outcome: Readmit"
author: "Steph Reynolds (Stephanie.Reynolds@ucsf.edu)"
date: "`r format(Sys.time(), '%b %d, %Y  %H:%M %p')`"
output:
  html_document:
    toc: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Background

## Project 
MINDSCAPE: Modeling of infectious network dynamics for surveillance, control and prevention enhancement

## Description
This file reads in the demographics, Covid staging, and ADI data (`dm_stg_adi_7.19.21.rds`) and fits a logistic regression model with readmit (`readmit`) as the outcome. Primary and secondary outcomes for full regression analysis can be found below. 

## Primary Outcome 

  * Morbidity and Mortality (`severe_stg`) -> All patients who reached a max stage of 6+ (severe) were recoded as severe_stg = 1

## Secondary Outcomes

  * Length of Stay (`LOS`) -> Length of stay that patients were hospitalized
    * Only patients who were hospitalized for >= 1 day - aka inpatients - were included in the regression analysis.
  * Readmit (`readmit`) --> Patients who were readmitted to hospital within 30(?) days

# Load required packages
```{r Load required packages, include=F}
library(tidyverse)
library(tidymodels)
library(performance) # to calculate model performance and check model assumptions
```

# Import and preview data
```{r Read in df}
df <- readRDS("/Users/sreynolds2/Documents/GitHub/MS-ADI/data/dm_stg_adi_7.19.21.rds")

head(df, n=10)
```

# Compare full and null models -- Readmit (`readmit`)
```{r Full and null models of LOS}
# FULL MODEL 
m_full_readmit <- glm(readmit ~ age + sex + smoking + cci_total + marital_status + language + insurance_type + pat_race + ethnicity + ADI + admission_stage + BMI, data = df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_full_readmit), confint.default(m_full_readmit, level=0.95)))

# Summarize regression output 
summary(m_full_readmit)

# Check model assumptions
check_model(m_full_readmit)

# Check model performance 
model_performance(m_full_readmit)

# NULL MODEL 
m_null_readmit <- glm(readmit ~ 1, data = df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_null_readmit), confint.default(m_null_readmit, level=0.95)))

# Summarize regression output 
summary(m_null_readmit)

# Check model performance 
model_performance(m_null_readmit)

```

# Generate simple logistic regressions for readmit

## Readmit ~ age
```{r Readmit ~ age}
m_readmit_age <- glm(readmit ~ age, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_age), confint.default(m_readmit_age, level=0.95)))

# Summarize regression output 
summary(m_readmit_age)

# Check model assumptions
check_model(m_readmit_age)

# Check model performance 
model_performance(m_readmit_age)
```

## Readmit ~ sex 
```{r Readmit ~ sex}
m_readmit_sex <- glm(readmit ~ sex, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_sex), confint.default(m_readmit_sex, level=0.95)))

# Summarize regression output 
summary(m_readmit_sex)

# Check model assumptions
check_model(m_readmit_sex)

# Check model performance 
model_performance(m_readmit_sex)
```

## Readmit ~ smoking
```{r Readmit ~ smoking}
m_readmit_smoke <- glm(readmit ~ smoking, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_smoke), confint.default(m_readmit_smoke, level=0.95)))

# Summarize regression output 
summary(m_readmit_smoke)

# Check model assumptions
check_model(m_readmit_smoke)

# Check model performance 
model_performance(m_readmit_smoke)
```

## Readmit ~ Charlston Comorbidity Index (CCI)
```{r Readmit ~ cci}
m_readmit_cci <- glm(readmit ~ cci_total, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_cci), confint.default(m_readmit_cci, level=0.95)))

# Summarize regression output 
summary(m_readmit_cci)

# Check model assumptions
check_model(m_readmit_cci)

# Check model performance 
model_performance(m_readmit_cci)
```

## Readmit ~ marital status
```{r Readmit ~ marital status}
m_readmit_marital <- glm(readmit ~ marital_status, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_marital), confint.default(m_readmit_marital, level=0.95)))

# Summarize regression output 
summary(m_readmit_marital)

# Check model assumptions
check_model(m_readmit_marital)

# Check model performance 
model_performance(m_readmit_marital)
```

## Readmit ~ language
```{r Readmit ~ language}
m_readmit_language <- glm(readmit ~ language, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_language), confint.default(m_readmit_language, level=0.95)))

# Summarize regression output 
summary(m_readmit_language)

# Check model assumptions
check_model(m_readmit_language)

# Check model performance 
model_performance(m_readmit_language)
```

## Readmit ~ insurance type
```{r Readmit ~ insurance}
m_readmit_insur <- glm(readmit ~ insurance_type, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_insur), confint.default(m_readmit_insur, level=0.95)))

# Summarize regression output 
summary(m_readmit_insur)

# Check model assumptions
check_model(m_readmit_insur)

# Check model performance 
model_performance(m_readmit_insur)
```

## Readmit ~ race
```{r Readmit ~ race}
m_readmit_race <- glm(readmit ~ pat_race, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_race), confint.default(m_readmit_race, level=0.95)))

# Summarize regression output 
summary(m_readmit_race)

# Check model assumptions
check_model(m_readmit_race)

# Check model performance 
model_performance(m_readmit_race)
```

## Readmit ~ ethnicity
```{r Readmit ~ ethnicity}
m_readmit_ethnicity <- glm(readmit ~ ethnicity, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_ethnicity), confint.default(m_readmit_ethnicity, level=0.95)))

# Summarize regression output 
summary(m_readmit_ethnicity)

# Check model assumptions
check_model(m_readmit_ethnicity)

# Check model performance 
model_performance(m_readmit_ethnicity)
```

## Readmit ~ Area Deprivation Index (ADI)
```{r Readmit ~ ADI}
m_readmit_adi <- glm(readmit ~ ADI, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_adi), confint.default(m_readmit_adi, level=0.95)))

# Summarize regression output 
summary(m_readmit_adi)

# Check model assumptions
check_model(m_readmit_adi)

# Check model performance 
model_performance(m_readmit_adi)
```

## Readmit ~ admission stage
```{r Readmit ~ admission stage}
m_readmit_adm_stg <- glm(readmit ~ admission_stage, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_adm_stg), confint.default(m_readmit_adm_stg, level=0.95)))

# Summarize regression output 
summary(m_readmit_adm_stg)

# Check model assumptions
check_model(m_readmit_adm_stg)

# Check model performance 
model_performance(m_readmit_adm_stg)
```

## Readmit ~ BMI
```{r Readmit ~ BMI}
m_readmit_BMI <- glm(readmit ~  BMI, df, family = 'binomial')

# Odds ratio of coefficients
exp(cbind("Odds ratio"=coef(m_readmit_BMI), confint.default(m_readmit_BMI, level=0.95)))

# Summarize regression output 
summary(m_readmit_BMI)

# Check model assumptions
check_model(m_readmit_BMI)

# Check model performance 
model_performance(m_readmit_BMI)
```
